move components to more organized locations
